+++ Monkey/Views/HeaderSegmentControl.h.m
@@ -86,7 +86,7 @@ UIFont *lightFont;
     }
     return self;
 }
--(void)swipeAction:(int)tag{
+-(void)swipeAction:(NSInteger)tag{
     UIColor *black=[UIColor blackColor];
     switch (tag) {
         case 101:
